# [Lua] Library improvements