Realized `view.lua#Offset()` is in a hot path and optimized it to avoid a bunch of needless lookups. I first cache the location to the vtable (instead of calculating it everytime) and inlined the unpacking function. 

This lead me to optimized most of the `view.lua` file to reduce lookups and use local functions as much as possible. I also hardcoded a lot of values that will never change, again to avoid lookups.

The profiling tests show almost a 2x improvement in traversing the flatbuffer data.

**Before**:

```
built 100 512-byte flatbuffers in 0.03sec: 3.67/msec, 1.79MB/sec
built 1000 512-byte flatbuffers in 0.22sec: 4.62/msec, 2.26MB/sec
built 10000 512-byte flatbuffers in 2.17sec: 4.61/msec, 2.25MB/sec
traversed 100 440-byte flatbuffers in 0.01sec: 15.10/msec, 6.34MB/sec
traversed 1000 440-byte flatbuffers in 0.07sec: 14.80/msec, 6.21MB/sec
traversed 10000 440-byte flatbuffers in 0.69sec: 14.40/msec, 6.04MB/sec
Number of test passed: 10 / 10 (100.00%)
```

**After**:

```
built 100 512-byte flatbuffers in 0.03sec: 2.95/msec, 1.44MB/sec
built 1000 512-byte flatbuffers in 0.20sec: 4.96/msec, 2.42MB/sec
built 10000 512-byte flatbuffers in 1.93sec: 5.17/msec, 2.53MB/sec
built 10000 512-byte flatbuffers in 1.52sec: 6.59/msec, 3.22MB/sec
traversed 100 440-byte flatbuffers in 0.00sec: 26.85/msec, 11.27MB/sec
traversed 1000 440-byte flatbuffers in 0.04sec: 26.30/msec, 11.03MB/sec
traversed 10000 440-byte flatbuffers in 0.36sec: 27.40/msec, 11.50MB/sec
Number of test passed: 11 / 11 (100.00%)

```