@aardappel 2019-11-14
Wow, that's impressive!
The hardcoded numbers are unfortunate, but I guess that's what is needed here.