@dbaileychess 2019-11-15
I'll actually but those hard coded numbers into local up-values in a follow up PR so we will have better flexibility without a hash lookup each time